get_subset moved to data_access | stronger tests
