+++ kale/loaddata/multi_domain.py
@@ -10,9 +10,8 @@ import numpy as np
 import torch.utils.data
 from sklearn.utils import check_random_state
 
-from kale.loaddata.dataset_access import DatasetAccess
+from kale.loaddata.dataset_access import DatasetAccess, get_class_subset
 from kale.loaddata.sampler import get_labels, MultiDataLoader, SamplingConfig
-from kale.utils.class_subset import get_class_subset
 
 
 class WeightingType(Enum):
